refactor(layer-metadata): move service/helper classes and utils to service folder
